Store comic location in a file instead of cookies
So, for some reason, using cookies seems to be a bit unreliable in chrome on my One Plus 5.
Since the main object of this software is to make it easier to read Darths and Droids piecemeal
I switch to using a file on the server. This also marks the switch from something I'm ok
showing to the internet at large, to something I need to put behind auth.

Also, this now uses https links for Darths and Droids, since there seems to be a working (if slower) https version
of the site, which, YAY! That saves me from having to write a messy scraper/rewriter/proxy to get this working on https.